Update to `BlockBuffer`
