+++ rand_isaac/src/lib.rs
@@ -31,7 +31,7 @@ pub mod isaac;
 pub mod isaac64;
 
 #[cfg(feature = "serde")]
-mod array_serde;
+mod serde_impl;
 
 pub use self::isaac::IsaacRng;
 pub use self::isaac64::Isaac64Rng;
